requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
websockets==10.4
